test: add extension to integration tests
