+++ extension/Cargo.toml
@@ -26,7 +26,6 @@ sp-runtime.workspace = true
 sp-std.workspace = true
 
 [dev-dependencies]
-contract-build.workspace = true
 env_logger.workspace = true
 pallet-balances.workspace = true
 pallet-contracts.workspace = true
