+++ integration-tests/Cargo.toml
@@ -8,7 +8,16 @@ repository.workspace = true
 
 [dev-dependencies]
 codec.workspace = true
-tracing-subscriber = { workspace = true, features = ["env-filter", "fmt", "std", "tracing-log"] }
+contract-build.workspace = true
+env_logger.workspace = true
+log.workspace = true
+scale-info.workspace = true
+tracing-subscriber = { workspace = true, features = [
+    "env-filter",
+    "fmt",
+    "std",
+    "tracing-log",
+] }
 
 # Substrate
 frame-support.workspace = true
@@ -88,7 +97,6 @@ try-runtime = [
     "asset-hub-paseo-runtime/try-runtime",
     "frame-support/try-runtime",
     "pallet-assets/try-runtime",
-    "pallet-balances/std",
     "pallet-message-queue/try-runtime",
     "paseo-runtime/try-runtime",
     "pop-runtime-devnet/try-runtime",
