+++ tests/extension/mod.rs
@@ -1,0 +1,2 @@
+pub mod mock;
+mod tests;
